requirements.txt
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.5.0
Pillow==11.0.0
python-dotenv==1.0.0
drf-yasg==1.21.7
django-filter==23.5
